backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
websockets>=11.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
httpx>=0.24.0
openai>=1.6.0
python-multipart>=0.0.6
websocket-client>=1.5.0
pytest>=7.3.1
pytest-asyncio>=0.21.0 